add pipeline var to skip openssh server installation

+++ vhdbuilder/packer/windows/configure-windows-vhd.ps1
@@ -492,13 +492,18 @@ function Install-ContainerD
 
 function Install-OpenSSH
 {
+    if ($env:INSTALL_OPEN_SSH_SERVER -eq 'true')
+    {
+        Write-Log "Not installing Windows OpenSSH Server as this is disabled in the pipeline"
+        return
+    }
+
     Write-Log "Installing OpenSSH Server"
 
     # Somehow openssh client got added to Windows 2019 base image.
     if ($env:WindowsSKU -Like '2019*')
     {
         Remove-WindowsCapability -Online -Name OpenSSH.Client~~~~0.0.1.0
-        Add-WindowsCapability -Online -Name OpenSSH.Client~~~~0.0.1.0
     }
 
     Add-WindowsCapability -Online -Name OpenSSH.Server~~~~0.0.1.0
